pyproject.toml
[project]
authors = [{name = "Peter Gagarinov", email = "4868370+pgagarinov@users.noreply.github.com"}]
name = "doc-sql-audit"
requires-python = ">= 3.11"
version = "0.1.0"
dependencies = [ "psycopg2-binary>=2.9.11,<3", "faker>=40.5.1,<41", "numpy>=2.4.2,<3", "pandas>=3.0.1,<4", "pytest>=8.0,<9", "claude-agent-sdk>=0.1.44,<0.2"]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
doc_sql_audit = { path = ".", editable = true }

[tool.pytest.ini_options]
markers = ["agent: tests using Claude Agent SDK (slow, uses Claude Max subscription)"]

[tool.pixi.tasks]
test = "pytest tests/ -v -m 'not agent'"
test-agent = "pytest tests/ -v -m agent"
test-all = "pytest tests/ -v"
